pyproject.toml
[project]
name = "episemic-core"
version = "0.1.0"
description = "A brain-inspired memory system for AI agents"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "typer (>=0.18.0,<0.19.0)",
    "rich (>=14.1.0,<15.0.0)",
    "qdrant-client (>=1.15.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "fastapi (>=0.116.2,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)"
]

[project.scripts]
episemic = "episemic_core.cli.main:app"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "ruff (>=0.13.1,<0.14.0)",
    "pre-commit (>=4.3.0,<5.0.0)"
]
